+++ frontend/src/pages/AccountDeleteComplete.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import styles from '../style/complete.module.css';
-import backgroundImage from '../assets/images/main.jpg';
+import backgroundImage from '../assets/images/backgroundAccountManage.jpg';
 import checkImage from '../assets/images/Check.png';
 
 const AccountDeleteComplete = () => {
