+++ frontend/src/style/AccountDeleteRequestPage.module.css
@@ -86,6 +86,13 @@
     color: black;
 }
 
+.tab:hover {
+    background-color: #bbb;
+    color: black;
+    transition: background-color 0.3s, color 0.3s;
+}
+
+
 .container {
     background-color: rgba(0, 0, 0, 0.85);
     border-radius: 20px;
