+++ frontend/src/style/AccountLookupIntro.module.css
@@ -55,7 +55,7 @@
 
 .centered {
     position: absolute;
-    top: 50%;
+    top: 40%;
     left: 50%;
     transform: translate(-50%, -50%);
     text-align: center;
