+++ frontend/src/style/RealNameVerificationPage.module.css
@@ -1,4 +1,3 @@
-/* src/styles/RealNameVerificationPage.module.css */
 .wrapper {
     position: relative;
     width: 100%;
@@ -21,92 +20,88 @@
     z-index: 0;
 }
 
-.header {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    padding: 25px 40px;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    z-index: 1001;
-}
-
-.logo {
-    font-family: 'GeostarFill', serif;
-    font-size: 2em;
-    font-weight: normal;
-    color: #fff;
-    cursor: pointer;
-    text-decoration: none;
-}
-
-.profileButton {
-    background: none;
-    color: #fff;
-    font-size: 0.9em;
-    cursor: pointer;
-    border: none;
-    border-radius: 5px;
-    white-space: nowrap;
-    z-index: 1002;
-}
-
-.contentBox {
+.innerContent {
     position: relative;
-    z-index: 1;
-    background: black;
+    z-index: 2;
+    width: 100%;
+    max-width: 1000px;
+    margin: 140px auto 0;
+    padding: 50px 60px;
+    background: rgba(0, 0, 0, 0.75);
+    backdrop-filter: blur(12px);
+    border-radius: 18px;
+    border: 1px solid rgba(255, 255, 255, 0.1);
     color: white;
-    width: 950px;
-    margin: 160px auto 0;
-    padding: 20px 50px;
-    border-radius: 30px;
-    border: 1px solid #d9d9d9;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    box-shadow: 0 20px 40px rgba(0, 0, 0, 0.4);
 }
 
 .title {
-    font-size: 40px;
+    font-size: 2.6rem;
     font-weight: 800;
-    margin-bottom: 20px;
+    margin-bottom: 0.8rem;
+    text-align: left;
 }
 
 .desc {
-    font-size: 18px;
-    font-weight: 500;
-    margin-bottom: 10px;
+    font-size: 1.1rem;
+    font-weight: 600;
+    margin-bottom: 0.6rem;
+    text-align: left;
+    color: #f5f5f5;
 }
 
 .notice {
-    font-size: 14px;
-    line-height: 22px;
-    color: #ccc;
-    margin-bottom: 10px;
+    font-size: 0.95rem;
+    color: #bbb;
+    line-height: 1.6;
+    margin-bottom: 2.2rem;
+    max-width: 820px;
 }
 
 .red {
     color: #ff5a5a;
+    font-weight: bold;
 }
 
-.formGroup {
+.formRow {
+    display: flex;
+    gap: 50px;
+    width: 100%;
+    align-items: flex-start;
+}
+
+.inputBox {
     display: flex;
     flex-direction: column;
-    gap: 5px;
+    width: 100%;
+    max-width: 300px;
 }
 
 .label {
-    font-size: 16px;
+    font-size: 1rem;
     font-weight: 600;
+    margin-bottom: 8px;
 }
 
 .input {
-    background: #222;
+    background: rgba(255, 255, 255, 0.08);
     color: white;
-    border: none;
+    border: 1px solid rgba(255, 255, 255, 0.2);
     border-radius: 10px;
-    padding: 10px 15px;
-    font-size: 16px;
-    width: 300px;
+    padding: 12px 16px;
+    font-size: 1rem;
+    outline: none;
+    transition: all 0.2s ease;
+    width: 100%;
+}
+
+.input:focus {
+    background: rgba(255, 255, 255, 0.1);
+    border-color: #fff;
+    box-shadow: 0 0 6px rgba(255, 255, 255, 0.3);
 }
 
 .ssnGroup {
@@ -117,30 +112,38 @@
 
 .hyphen {
     color: white;
-    font-size: 18px;
+    font-size: 20px;
+    user-select: none;
 }
 
-.buttonContainer {
-    position: relative;
+.buttonWrapper {
     width: 100%;
+    margin-top: 40px;
     display: flex;
-    justify-content: center;
-    margin-top: 20px;
-    z-index: 2;
+    justify-content: flex-end;
 }
 
 .button {
-    padding: 14px 40px;
-    background: #d9d9d9;
-    color: black;
-    font-size: 18px;
-    border: none;
-    border-radius: 10px;
-    cursor: pointer;
+    padding: 14px 48px;
+    background-color: #fff;
+    color: #000;
+    font-size: 1rem;
     font-weight: bold;
+    border: none;
+    border-radius: 8px;
+    cursor: pointer;
+    transition: all 0.2s ease;
+    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.25);
 }
 
 .button:disabled {
-    background-color: #888;
+    background-color: #777;
+    color: #3b3b3b;
     cursor: not-allowed;
+    box-shadow: none;
+}
+
+.button:hover:enabled {
+    background-color: #eee;
+    transform: translateY(-1px);
 }
